Added support for production units
Profit when raw production is transformed into artisan goods now takes into account the number of production units the player has, so that profit is reduced for high-yield, low-value crops.
This is a first step, as this early production simulation does not take into account production units productivity (depending essentially on crops, but not only), thus simulating 1 crop transformation/production unit/harvest. This skews the simulation in favor of high-yield crops, as the production unit might normally not be available at the time of the next harvest.
Production units productivity shall be taken into account in the future.